// apps/web/src/components/features/wallet/multiButton/types.ts
export enum WalletButtonLabel {
    CHANGE_WALLET = "changeWallet",
    CONNECTING = "connecting",
    COPY_ADDRESS = "copyAddress",
    COPIED = "copied",
    DISCONNECT = "disconnect",
    NO_WALLET = "noWallet",
    HAS_WALLET = "hasWallet",
    AUTHENTICATION_REQUIRED = "authenticationRequired",
}

export enum WalletButtonState {
    CONNECTING = "connecting",
    CONNECTED = "connected",
    DISCONNECTING = "disconnecting",
    HAS_WALLET = "has-wallet",
    NO_WALLET = "no-wallet",
    AUTHENTICATION_REQUIRED = "authenticationRequired",
}